% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate DNA into protein}
\usage{
translate(dna)
}
\arguments{
\item{dna}{a vector of numeric values}
}
\value{
protein sequence
}
\description{
Translate DNA into protein
}
\examples{
dna <- "ATCGCTGGA"
translate(dna)
}
