\name{transr-package}
\alias{transr-package}
\alias{transr}
\docType{package}
\title{
  Applying normalizing and variance stabilizing transformations to data
}
\description{
  This package contains several normalising and variance stabilizing transformations, methods to infer transformation parameters and functions to apply the forward and inverse transformation.
}
\details{
\tabular{ll}{ Package: \tab transr\cr Type: \tab Package\cr License: \tab
TBA\cr Date: \tab	 2019-09-16\cr  }
}
\author{
Andrew Schepen, David Robertson, and Jean-Michel Perraud.

Maintainer: Jean-Michel Perraud <per202@csiro.au>
}
\references{
  Box, G. E. P. and Cox, D. R. (1964) An analysis of transformations. \emph{Journal Royal Statistical Society: Series B} \bold{26} 211--246. https://doi.org/10.1111/j.2517-6161.1964.tb00553.x
  
  Wang, Q., D. L. Shrestha, D. Robertson and P. Pokhrel (2012). A log-sinh transformation for data normalization and variance stabilization. Water Resources Research 48(5). http://dx.doi.org/10.1029/2011WR010973.
  
  Yeo, I. K. and R. A. Johnson (2000). A new family of power transformations to improve normality or symmetry. Biometrika 87(4): 954-959. http://dx.doi.org/10.1093/biomet/87.4.954.
}
\keyword{ package }
\seealso{
  \code{\link{LogSinh}},  \code{\link{BoxCox}},  \code{\link{YeoJohnson}}
}

